requirements.txt
altgraph==0.17.4
appdirs==1.4.4
beautifulsoup4==4.12.2
BlurWindow==1.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
frozendict==2.3.8
html5lib==1.1
idna==3.7
lxml==4.9.3
multitasking==0.0.11
numpy==1.26.3
packaging==23.2
pandas==2.1.1
peewee==3.16.3
pefile==2023.2.7
Pillow==10.3.0
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
pyqtgraph==0.13.3
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32-ctypes==0.2.2
requests==2.31.0
six==1.16.0
soupsieve==2.5
tzdata==2023.3
urllib3==2.2.1
webencodings==0.5.1
yfinance==0.2.36
pyobjc==10.1; sys_platform == 'darwin' # MacOS only
#PyQt5 conflicts with pyqt6
# PyQt5==5.15.10 
# to install universal2 ver, need to install with python3 -m pip inst... (to make sure universal2 python ver is used)